create brdg_base.py, move models function there from brdg_02r13 (must tweak vmd_base default for cli)
