+++ src/pyairios/models/vmd_base.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import datetime
 from dataclasses import dataclass, field
 
-from pyairios.constants import VMDCapabilities, ValueStatusSource, ValueStatusFlags
+from pyairios.constants import ValueStatusFlags, ValueStatusSource, VMDCapabilities
 from pyairios.device import AiriosDevice
 from pyairios.registers import (
     RegisterAddress,
@@ -19,9 +19,9 @@ class VMDPresetFansSpeeds:
     """Preset fan speeds."""
 
     # this must load from vmd_base to prevent None error
-    exhaust_fan_speed: Result[int] = field(default=Result(-1))
+    exhaust_fan_speed: Result[int] = field()
     """Exhaust fan speed (%)"""
-    supply_fan_speed: Result[int] = field(default=Result(-1))
+    supply_fan_speed: Result[int] = field()
     """Supply fan speed (%)"""
 
     def __post_init__(self):
